requirements.txt
beautifulsoup4>=4.12.2
requests>=2.31.0
playwright
fastapi[standard]
sqlalchemy
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
uvicorn